+++ src/main/java/com/harium/etyl/geometry/curve/CubicBezier3.java
@@ -33,7 +33,7 @@ public class CubicBezier3 extends Curve3 {
         return p3;
     }
 
-    public void translate(float x, float y, float z) {
+    public void translate(double x, double y, double z) {
         super.translate(x, y, z);
 
         p2.x += x;
