+++ src/main/java/com/harium/etyl/geometry/curve/QuadraticBezier.java
@@ -34,7 +34,7 @@ public class QuadraticBezier extends Curve {
         return p2;
     }
 
-    public void translate(float x, float y) {
+    public void translate(double x, double y) {
         super.translate(x, y);
 
         p2.x += x;
